Fix meter API and config
1. Now meter API works.
2. Added missing `ai_model` to config.
3. Commented Tariff in meter for now.

+++ src/api/views/meter.py
@@ -15,7 +15,7 @@ class Meter(BaseModel):
     meter_id: int
     facility_id: int
     name:  Optional[str]
-    rating:  Optional[int]
+    rating:  Optional[float]
     address:  Optional[str]
 
     meter_details : MeterDetail
